+++ src/visualizers/graph/clientGraph.tsx
@@ -1,8 +1,0 @@
-'use client'
-
-import { clientOnly } from '@/visualizers/client/clientOnly'
-import GraphVisualizer from './graphVisualizer'
-
-const ClientGraph = clientOnly(GraphVisualizer)
-
-export default ClientGraph