+++ src/visualizers/sorting/clientSorting.tsx
@@ -1,8 +1,0 @@
-'use client'
-
-import { clientOnly } from '@/visualizers/client/clientOnly'
-import SortingVisualizer from './sortingVisualizer'
-
-const ClientSorting = clientOnly(SortingVisualizer)
-
-export default ClientSorting